requirements.txt
azure-ai-inference
mistralai
OpenAI
requests
